Настраивает окружение
