+++ src/components/events-list.js
@@ -1,0 +1,8 @@
+const getEventsListTemplate = () => {
+  return (
+    `<ul class="trip-days">
+     </ul>`
+  );
+};
+
+export { getEventsListTemplate};
