+++ src/components/menu.js
@@ -1,0 +1,10 @@
+const getMenuTemplate = () => {
+  return (
+    `<nav class="trip-controls__trip-tabs  trip-tabs">
+			<a class="trip-tabs__btn" href="#">Table</a>
+		 	<a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Stats</a>
+		</nav>`
+  );
+};
+
+export {getMenuTemplate};
